requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
distro==1.9.0
duckduckgo_search==7.1.1
exa_py==1.7.1
fastapi==0.115.6
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
groq==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.8.2
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.1
openai==1.58.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
phidata==2.7.5
pillow==11.0.0
platformdirs==4.3.6
primp==0.9.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rsa==4.9
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tinysegmenter==0.3
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
yfinance==0.2.51
